Added method to calculate the costheta and phi for viewing angles given the viewing angle numbers  (#26)
* Removed changes I didnt want in the pull request

* Added line to fix PEP 8 issue

* simplified method for calculating costheta and phi bins to one Luke suggested


* Added test to produce error is MABINS is not equal to 100

+++ artistools/__init__.py
@@ -391,7 +391,7 @@ def get_vpkt_data(modelpath):
     with open(filename, 'r') as vpkt:
         vpkt_data['nobservations'] = int(vpkt.readline())
         vpkt_data['cos_theta'] = [float(x) for x in vpkt.readline().split()]
-        vpkt_data['phi'] = [int(x) for x in vpkt.readline().split()]
+        vpkt_data['phi'] = [float(x) for x in vpkt.readline().split()]
         line4 = vpkt.readline()
         _, vpkt_data['initial_time'], vpkt_data['final_time'] = [int(x) for x in vpkt.readline().split()]
         return vpkt_data
